jobs/pyproject.toml
[tool.poetry]
name = "jobs"
version = "0.1.0"
description = "Jobs microservice for BadgerDoc project"
authors = ["BadgerDoc team"]

[tool.poetry.dependencies]
python = "^3.8.0"
alembic = "^1.7.4"
SQLAlchemy = "1.3.24"
sqlalchemy_utils = "^0.38.3"
fastapi = "^0.68.1"
python-dotenv = "^0.19.0"
uvicorn = "^0.15.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"
types-requests = "^2.25.11"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "*"
mypy = "^0.910"
pylint = "^2.11.1"
freezegun = "^1.1.0"
pytest-cov = "*"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
